% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myocard.R
\docType{data}
\name{myocard}
\alias{myocard}
\title{Myocarditis intercross data}
\format{An object of class \code{cross}.  See \code{\link[qtl:read.cross]{qtl::read.cross()}}
for details.}
\source{
Noel R. Rose, Department of Pathology, Johns Hopkins University,
\email{nrrose@biostat.wisc.edu}
}
\usage{
myocard
}
\description{
Data from a mouse intercross experiment on myocarditis.
}
\details{
An intercross between the H-2s congenic mice A.SW and B10.S, with 296
individuals (including both sexes). The mice were injected with purified
murine cardiac myosin, and the area of infiltrated myocardium in heart
sections was measured.  The phenotype is the percent myocarditis.
}
\examples{

data(myocard)
library(qtl)
plot(myocard)

}
\references{
Guler, M. L, Ligons, D. L., Wang, Y., Bianco, M., Broman, K. W.
and Rose, N. R. (2005) Two autoimmune diabetes loci influencing T cell
apoptosis control susceptibility to experimental autoimmune myocarditis.
\emph{J. Immunol.} \strong{174}: 2167--2173.
}
\seealso{
\code{\link[=gutlength]{gutlength()}}, \code{\link[=iron]{iron()}}, \code{\link[=nf1]{nf1()}}
\code{\link[=ovar]{ovar()}}, \code{\link[=trout]{trout()}}
}
\keyword{datasets}
